% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform_utils.R
\name{get_download_dir}
\alias{get_download_dir}
\title{Get default download directory that works across platforms}
\usage{
get_download_dir()
}
\value{
Path to the default download directory
}
\description{
Get default download directory that works across platforms
}
\keyword{internal}
