% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ces-package.R
\docType{package}
\name{ces-package}
\alias{ces}
\alias{ces-package}
\title{ces: Access to Canadian Election Study Data}
\description{
Provides tools to easily access and analyze Canadian Election Study data. The package simplifies the process of downloading, cleaning, and using CES datasets for political science research and analysis. The Canadian Election Study (CES) has been conducted during federal elections since 1965, surveying Canadians on their political preferences, engagement, and demographics. Data is accessed from the Borealis Data repository, which serves as the official host for CES datasets. This package is not officially affiliated with the Canadian Election Study or Borealis Data, and users should cite the original data sources in their work.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/laurenceomfoisy/ces}
  \item Report bugs at \url{https://github.com/laurenceomfoisy/ces/issues}
}

}
\author{
\strong{Maintainer}: Laurence-Olivier M. Foisy \email{mail@mfoisy.com} (\href{https://orcid.org/0009-0004-7505-9477}{ORCID})

}
\keyword{internal}
